+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 ### Features Added
-- Add CLI option `-Xkey-manager-skip-keystore-storage` to prevent web3signer from writing keystores imported via the keymanager API to disk. This is *not* safe for operators who do not backup keystore.
+- Add CLI option `-Xkey-manager-skip-keystore-storage` to eth2 subcommand preventing web3signer from writing keystores imported via the keymanager API to disk. This is *not* safe for operators who do not backup keystore.
 
 ## 25.9.1
 ### Breaking changes